+++ source/map3d/plugins/drawPolygon/polygonproperties.h
@@ -84,7 +84,7 @@ public:
     void setFillColor(const QColor &color) override;
     void setStrokeWidth(const double &opacity) override;
     void setStroke(const QColor &color) override;
-    void setClamp(const double &clamp) override;
+    void setClamp(const int &clamp) override;
     void setHeight(const double &height) override;
